chess/requirements.txt
channels==3.0.4
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
axios==0.4.0
langchain==0.2.11
langchain-experimental==0.0.63
langchain-openai==0.1.19
langgraph==0.1.15
chess==1.10.0
